fix: combine emoji logic
