+++ package.json
@@ -35,7 +35,6 @@
     "eslint": "^9",
     "eslint-config-next": "15.3.2",
     "tailwindcss": "^4",
-    "ts-node": "^10.9.2",
     "typescript": "^5.8.3"
   }
 }
